pose-classifier/requirements.txt
viam-sdk~=0.43.0
ultralytics~=8.3.94
scikit-learn~=1.6.0
numpy~=1.24.0
joblib~=1.3.0
twilio~=8.2.0